+++ src/lib/msal-client.js
